requirements.txt
easyocr==1.7.1
numpy==2.0.1
opencv_python==4.10.0.82
pandas==2.2.2
Pillow==10.4.0
python_Levenshtein==0.25.1